Create RSVP's Model, Add Agenda to Meeting Model
This commit creates the initial model for RSVP's. Each RSVP will belong to a meeting and a user.
Additionally, this commit adds an agenda (text field) to a meeting. In a future feature, an agenda
would most likely be pulled out into it's own model. This commit also adds a "displayable" boolean
on meetings for eventually deciding whether or not the agenda will be shown on it's public page.